requirements.txt
gunicorn==21.2.0
Flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.1.0
PyJWT==2.10.1
